requirements.txt
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
cffi==1.16.0
cryptography==42.0.2
discord.py==2.3.2
frozenlist==1.4.1
html5lib==1.1
idna==3.6
multidict==6.0.5
pycparser==2.21
PyMySQL==1.1.0
six==1.16.0
soupsieve==2.5
table2ascii==1.1.3
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
